Changed C source files such that they will compile under g++ without errors or warnings: - in functions gsProverInit and gsNextstateInit, the variable RewriteStrategy of   type int is replaced by strat of type RewriteStrategy, since this is needed   for a correctly typed call to gsRewriteInit - fixed implicit casting of bool to ATbool by:   + changing the return types of functions from ATbool to bool   + replacing return values ATtrue/ATfalse by true/false in these functions   + replacing true/false to ATtrue/ATfalse in the creation of AFun's - added forgotten type casts - solved const / non-const mismatch in the use of strings - renamed a variable because it was declared twice - renamed the variable "new" because it is reserved in C++
Note that I also tried to compile the SVC and ATerm library with g++. The SVC
library gave only a few warnings concerning illegal (but probably innocent)
type casts, but with the ATerm library all hell broke loose: the errors are
disturbing enough to require immediate attention.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@759 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ src/gslowlevel.c
@@ -438,11 +438,11 @@ char *gsStringDub(const char *n, const int inc)
 int NrOfChars(int n)
 {
   if (n > 0)
-    return ceil(log10((double) n));
+    return (int) ceil(log10((double) n));
   else if (n == 0)
     return 1;
   else //n < 0
-    return ceil(log10((double) abs(n))) + 1;
+    return (int) ceil(log10((double) abs(n))) + 1;
 }
 
 #ifdef __cplusplus
